src/requirements.txt
fastapi
torch==1.10.0
torchvision==0.11.1
numpy
pillow
prometheus-client
prometheus-fastapi-instrumentator
psutil